requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.12
langchain-community==0.0.28

# Vector Database & Embeddings
chromadb==0.4.24
sentence-transformers==2.5.1
huggingface-hub==0.21.4

# Document Loaders
pypdf==4.1.0
unstructured[md]==0.12.5

# LLM Integration
ollama==0.1.7
python-dotenv==1.0.1
